requirements.txt
pyrogram==2.0.106
aiohttp==3.7.4.post0
requests==2.26.0
python-dotenv==0.21.0
